Backend Completado
El desarrollo del backend ha finalizado de momento. El unico inconveniente es la visualizacion en swagger

+++ src/main/java/com/example/BodegaTrack/BodegaTrack/grocer/model/Grocer.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class Grocer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;;
+    private Long id;;
 
     @Column(name="name", nullable = false)
     private String name;
